feat: Add ArcanaFlow template and Solana Mobile Candy Machine mint sample
- Introduced a minimal ArcanaFlow template demonstrating frame batching, ALT proposal building, and v0 compilation.
- Implemented a Solana Mobile app for minting using Candy Machine v3, showcasing wallet connection and minting flow.
- Updated Gradle wrapper to version 8.10.2 and added necessary build scripts for both samples.
- Created README files for both samples to guide users through setup and execution.
- Added AndroidManifest and layout files for the mobile minting app.
- Included Kotlin code for main activity handling wallet connections and minting operations.

+++ artemis-gaming/src/main/kotlin/com/selenus/artemis/gaming/ComputeBudgetPresets.kt
@@ -18,8 +18,8 @@ object ComputeBudgetPresets {
 
   fun setComputeUnitLimit(units: Int): Instruction = ComputeBudgetProgram.setComputeUnitLimit(units)
 
-  fun setComputeUnitPrice(microLamports: Int): Instruction =
-    ComputeBudgetProgram.setComputeUnitPrice(microLamports.toLong())
+  fun setComputeUnitPrice(microLamports: Long): Instruction =
+    ComputeBudgetProgram.setComputeUnitPrice(microLamports)
 
   fun preset(tier: Tier): List<Instruction> {
     return listOf(
